# Inventario/API/Controllers/Producto/ProductDto.cs
namespace API.Controllers.Producto
{
    public class ProductDto
    {
        public string Nombre { get; set; }
        public string Descripcion { get; set; }
        public decimal Precio { get; set; }
        public int Stock { get; set; }
        public string Categoria { get; set; }

    }
}